+++ src/VMFirst.Test/IViewAwareViewModelTest.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading;
+﻿using System;
+using System.Threading;
 using System.Windows;
 using Moq;
 using NUnit.Framework;
@@ -25,12 +26,15 @@ namespace VMFirst.Test
 		[Apartment(ApartmentState.STA)]
 		public void ViewAwareViewModelHelper_Direct_Property_Injection_Succeeds()
 		{
+			// Arrange
 			var view = new Window();
 			var viewModel = new ViewAwareViewModelWithSetableProperty();
-			var setupCallback = ViewAwareViewModelHelper.CreateViewModelSetupCallback();
-			
+			Action<object, FrameworkElement> setupCallback = ViewAwareViewModelHelper.Callback;
+
+			// Act
 			setupCallback.Invoke(viewModel, view);
 
+			// Assert
 			Assert.That(viewModel.View, Is.EqualTo(view));
 		}
 
@@ -38,22 +42,28 @@ namespace VMFirst.Test
 		[Apartment(ApartmentState.STA)]
 		public void ViewAwareViewModelHelper_Indirect_Property_Injection_Succeeds()
 		{
+			// Arrange
 			var view = new Window();
 			var viewModel = new ViewAwareViewModelWithAutoProperty();
-			var setupCallback = ViewAwareViewModelHelper.CreateViewModelSetupCallback();
+			Action<object, FrameworkElement> setupCallback = ViewAwareViewModelHelper.Callback;
 
+			// Act
 			setupCallback.Invoke(viewModel, view);
 
+			// Assert
 			Assert.That(viewModel.View, Is.EqualTo(view));
 		}
 		[Test]
 		public void ViewAwareViewModelHelper_Property_Injection_Fails()
 		{
+			// Arrange
 			var viewModel = new Mock<IViewAwareViewModel>().Object; // Mocked object does not have a setter or a backing field.
-			var setupCallback = ViewAwareViewModelHelper.CreateViewModelSetupCallback();
+			Action<object, FrameworkElement> setupCallback = ViewAwareViewModelHelper.Callback;
 
+			// Act
 			setupCallback.Invoke(viewModel, null);
 
+			// Assert
 			Assert.That(viewModel.View, Is.Null);
 		}
 	}